Ne plus tronquer les qsizetype de Qt 6
Qt 6 fait passer size(), count(), indexOf() et lastIndexOf() de int à qsizetype.
Vingt endroits en ramenaient le résultat à un int.

Ces vingt-là ne produisaient aucun avertissement, et pas parce que MSVC
l'ignore : Qt désactive lui-même le C4244 pour toute unité de compilation qui
inclut un de ses en-têtes, dans le bloc QT_NO_WARNINGS de qcompilerdetection.h.
Un cl /W3 sur un cas équivalent écrit à la main le signale très bien. C'est ce
qui rend l'affaire piégeuse : la compilation était déjà muette à /W3 comme à /W4,
et il fallait définir QT_CC_WARNINGS pour les voir — ce qui sort aussi 4844
C4251 venant des en-têtes de Qt, donc ce n'est pas une option à laisser active.

MIGRATION-QT6.md ne mentionne pas qsizetype du tout, et l'estimation qui avait
servi à planifier ce commit était fausse sur deux points : ces conversions ne
sont pas visibles à la compilation ordinaire, et qMin(5, size()) compile très
bien, Qt 6 ayant ajouté une surcharge à deux types qui promeut ses arguments.

Le partage suit la règle habituelle : qsizetype là où c'est le type naturel — un
décalage dans une chaîne, un indice dans une liste — et cast explicite là où une
API veut vraiment un int, celles de Qt comme setCursorPosition ou setCurrentIndex
et les trois fonctions de favoris, dont l'argument est un numéro d'emplacement
borné par une poignée de valeurs.

Le C4244 restant, un double vers int dans manageShortcutWindow, est antérieur au
portage et sans rapport avec Qt 6 : la troncature y est voulue, elle est
simplement rendue explicite.

Vérifié : plus aucun avertissement dans les sources du dépôt, QT_CC_WARNINGS
compris.

+++ respawnIrc/imageDownloadTool.cpp
@@ -151,7 +151,7 @@ QString imageDownloadToolClass::getPathOfTmpDir()
 
 int imageDownloadToolClass::getNumberOfDownloadRemaining()
 {
-    return listOfImagesUrlNeedDownload.size();
+    return static_cast<int>(listOfImagesUrlNeedDownload.size());
 }
 
 bool imageDownloadToolClass::checkIfImageUrlExist(QString imageUrl, const imageDownloadRuleStruct& thisRule, QString ruleName)
